pyproject.toml
[project]
name = "cinema"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django>=5.2.5",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "environs[django]>=14.3.0",
    "ipdb>=0.13.13",
    "markdown>=3.8.2",
    "psycopg[binary]>=3.2.9",
    "requests>=2.32.5",
    "whitenoise>=6.9.0",
]

[tool.black]
target-version = ["py313"]

[tool.coverage.run]
omit = [
    "*/admin.py",
    "*/manage.py",
    "*/migrations/*",
    "*/tests/*",
    "conftest.py",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
addopts = "--cov --nomigrations --reuse-db"
norecursedirs = ".git* frontend media static templates"
python_files = "test_*.py"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".github",
    ".hg",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "migrations",
]
# Same as Black.
line-length = 80
# Assume Python 3.13.
target-version = "py313"

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F"]
ignore = ["E501", "E741"]                # temporary
per-file-ignores = {}
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
unfixable = []

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "model-bakery>=1.20.5",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
]

[project.optional-dependencies]
lint = ["ruff>=0.12.9"]
